Added isOpen method to Dialect
Added support for optional add... method in entity object, for to support custom
collection loading (to use, implement addObject(Object o) instead of setObjects,
and addObject will be called for each object in loaded in the collection).  Note
that the addObject must then add the object to a collection, if that is the
desired behavior.